+++ R/point_interval.R
@@ -145,7 +145,6 @@ globalVariables(c("y", "ymin", "ymax"))
 #'
 #' @importFrom purrr map_dfr map map2 discard map_dbl map_lgl iwalk
 #' @importFrom dplyr do bind_cols group_vars summarise_at
-#' @importFrom stringi stri_startswith_fixed
 #' @importFrom tidyr unnest
 #' @importFrom rlang set_names quos quos_auto_name eval_tidy as_quosure
 #' @export
@@ -338,12 +337,24 @@ hdi = function(x, .width = .95, .prob, na.rm = FALSE) {
 
 #' @export
 #' @rdname point_interval
+#' @importFrom rlang is_integerish
 Mode = function(x, na.rm = FALSE) {
-  if (!na.rm && any(is.na(x))) {
+  if (na.rm) {
+    x = x[!is.na(x)]
+  }
+  else if (any(is.na(x))) {
     return(NA_real_)
   }
 
-  LaplacesDemon::Mode(x)
+  if (is_integerish(x)) {
+    # for the discrete case, based on https://stackoverflow.com/a/8189441
+    ux = unique(x)
+    ux[which.max(tabulate(match(x, ux)))]
+  } else {
+    # for the continuous case
+    d = density(x)
+    d$x[which.max(d$y)]
+  }
 }
 
 #' @export