+++ tests/testthat/test.point_interval.R
@@ -214,7 +214,7 @@ test_that("multiple-response intervals work", {
     summarise(
       .lower = list(hdi(x, .width = .5)[, 1]),
       .upper = list(hdi(x, .width = .5)[, 2]),
-      x = LaplacesDemon::Mode(x),
+      x = Mode(x),
       .width = .5,
       .point = "mode",
       .interval = "hdi"
@@ -227,6 +227,24 @@ test_that("multiple-response intervals work", {
 test_that("point_interval errors if there are no columns to summarise", {
   expect_error(median_hdi(data.frame()),
     "No columns found to calculate point and interval summaries for\\.")
+})
+
+test_that("point_interval works on vectors", {
+  set.seed(1234)
+  x = rnorm(100, mean = 5)
+
+  ref = data.frame(
+    y = mean(x),
+    ymin = as.vector(quantile(x, probs = .025)),
+    ymax = as.vector(quantile(x, probs = .975)),
+    .width = .95,
+    .point = "mean",
+    .interval = "qi",
+    stringsAsFactors = FALSE
+  )
+
+  expect_equal(mean_qi(x), ref)
+  expect_equal(mean_qih(x), rename(ref, x = y, xmin = ymin, xmax = ymax))
 })
 
 test_that("various point summaries and intervals give correct numbers", {